fix: move to utils
